Angel's Compatibility Fixes (#316)
* Fixes Angel's Smelting Compatibility

Icon sizes are different than Krastorio's default usage. 
`steel-plate` also doesn't have an icons[2] so that would trip during initialization and fail causing a crash.

* Fixes Angel's Refining Compat

This is a minor change, but the icon now displays properly in game. If using size 32, the icon in the top left corner of the recipe disappears.

* Fixes Angel's Petrochem Compatibility

Disables resin-{n}, angels-flare-stack, & rubber technologies instead of setting them to nil. Startup seemingly one or more of those recipes were referenced later causing a crash.

Removes utility-science-packs from the research for rocket fuel when angel's petro is in. This otherwise causes a research dead end that is a loop. Without this: Rocket fuel needs utility science packs to research, and utility science packs need rocket fuel to craft.